refactor: move DragAndDropModeSelector file
