+++ src/components/layout/banner/actions/DragAndDropModeSelector.vue
